Rename visualization to visualization frame to resolve linting warning
